services/processor/requirements.txt
# SysArmor Processor - Python Dependencies
# PyFlink 实时威胁检测处理器核心依赖包

#==============================================================================
# Core PyFlink Dependencies - 核心 PyFlink 依赖
#==============================================================================

# Apache Flink Python API
apache-flink==1.18.1
apache-flink-libraries==1.18.1

#==============================================================================
# Data Processing - 数据处理
#==============================================================================

# Kafka 客户端
kafka-python==2.0.2

# OpenSearch 客户端
opensearch-py==2.3.1

# JSON 处理
ujson==5.8.0

# 日期时间处理
python-dateutil==2.8.2

# 正则表达式增强
regex==2023.8.8

#==============================================================================
# Configuration & Logging - 配置和日志
#==============================================================================

# 配置管理
python-dotenv==1.0.0
PyYAML==6.0.1

# 环境变量管理
environs==9.5.0

#==============================================================================
# Utilities - 实用工具
#==============================================================================

# 类型检查
typing-extensions==4.7.1

# 网络工具
requests==2.31.0

# 系统信息
psutil==5.9.5
